refactor: :recycle: selector 개념 재정의 및 일부 코드 수정
